feat(cli): handle imports in the migration generator
Previously, `use` statements had to be added manually after generating a
migration. This commit adds a simple solution that scans the file for
top-level `use`s, `struct`s and `const`s so that they are imported
automatically.

In addition to that, this adds some tests to the migration generator and
some useful options to the CLI (--app-name and --output-dir to override
the migration's app name and generator's output directory).